fixed route-protection to eliminate errors and handle redirection properly
