+++ frontend/src/App.js
@@ -4,6 +4,7 @@ import Public from './components/Public';
 import DashLayout from './components/DashLayout';
 import EmployeesList from './features/employee/EmployeesList';
 import Login from './features/auth/Login';
+import Unauthorized from './components/Unauthorized';
 import RequireAuth from './features/auth/RequireAuth';
 import Home from './features/auth/Home';
 import { ROLES } from './config/roles';
@@ -15,6 +16,7 @@ function App() {
 				<Route index element={<Public />}/>
 
 				<Route path='login' element={<Login/>}/>
+                <Route path='unauthorized' element={<Unauthorized/>}/>
 
 				<Route element={<RequireAuth allowedRoles={[ROLES.Employee, ROLES.Manager, ROLES.Admin]} />} >
 					<Route path="dash" element={<DashLayout />} >
