+++ frontend/src/components/Unauthorized.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+const Unauthorized = () => {
+    return (
+        <div>
+            <h1>Unauthorized</h1>
+            <p>You do not have permission to view this page.</p>
+        </div>
+    );
+};
+
+export default Unauthorized;
