+++ frontend/src/features/auth/authSlice.js
@@ -18,4 +18,4 @@ export const { setCredentials, logOut } = authSlice;
 
 export default authSlice.reducer;
 
-export const selectCurrentToken = (state) => state.auth.token;
+export const selectCurrentToken = (state) => state.auth?.token;
